Report/Remove unused referred vars (#33)
* feat: remove unused-refers

Updates carve to handle clj-kondo's :findings, specifically the
:unused-referred-var type. These vars are included in the report or
removed from the code.

Includes an update to the clj-kondo :sha, so that the findings include
the name and namespace of the unused-ref. This is to support showing the
vars in the printed report.

* fix: exclude refers from report

Refers don't need to be included in the report - clj-kondo is a better
fit and does this already. This updates the tests to reflect that.