feat: better interactive refer-removal message (#34)
* feat: better interactive refer-removal message

Updates the clj-kondo dep to the latest, to be sure refer-removal runs
smoothly without the included ns and name in clj-kondo's findings.

The context presented during interactive mode for refers was weak - only
showing the name of the var, no namespace or surrounding lines. This
commit lifts babashka's error-context function and uses it as a
fallback for presenting context when the symbol at the location is not
known. I'm hopeful this fallback to surrouding lines will be immediately
applicable if more use-cases popup (using other clj-kondo findings, for
example.)

The interact function has been extended to prevent suggesting/adding a
nil symbol to the carve_ignore.

* test: coverage for interactive mode and refer-removal context

Adds a couple tests for interactive mode, ensuring that
removing/skipping vars works as expected.

Tests the output for the context returned when an unused refer is up for
deletion.